updated field type

+++ db/migrations/000011_add_post_reward.sql
@@ -1,6 +1,6 @@
 -- +migrate Up
 ALTER TABLE posts
-  ADD COLUMN total_upvote_reward int NOT NULL DEFAULT 0;
+  ADD COLUMN total_upvote_reward bigint NOT NULL DEFAULT 0;
 
 -- +migrate Down
 ALTER TABLE posts
